Refactor src files in src/Utils
